pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.3.0"
description = ""
authors = ["Anna Bystrova <annbystr@gmail.com>"]
readme = "README.md"
repository = "https://github.com/Aluwian/python-project-52"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
python-dotenv = "^1.0.0"
dj-database-url = "^2.0.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.7"
django-bootstrap5 = "^23.3"
whitenoise = "^6.5.0"
django-filter = "^23.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
coverage = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
